moving Fedora JMS consuming to WSUDOR_Indexer
