+++ WSUDOR_Indexer/__init__.py
@@ -1,83 +1,7 @@
-import xmltodict, json
-import localConfig
-
-# index to Solr
-from WSUDOR_Manager.actions.solrIndexer import solrIndexer
-from WSUDOR_Manager.actions.pruneSolr import pruneSolr_worker
-
-# handles events in Fedora Commons as reported by JSM
-class WSUDOR_Indexer(object):
-
-	def __init__(self, frame):
-
-		self.frame = frame
-		self.headers = frame.headers
-		self.methodName = self.headers['methodName']
-		self.pid = self.headers['pid']
-		self.ds = False
-		self.body = frame.body
-		self.parsed_body = xmltodict.parse(self.body)
-		self.title = self.parsed_body['entry']['title']['#text']
-		self.categories = self.parsed_body['entry']['category']
-
-		# method type
-		if self.methodName.startswith('add'):
-			self.methodType = 'add'
-		if self.methodName.startswith('modify'):
-			self.methodType = 'modify'
-		elif self.methodName.startswith('purge'):
-			self.methodType = 'purge'
-		else:
-			self.methodType = False
-
-
-	def act(self):
-
-		print "Fedora message: %s, consumed for: %s" % (self.methodName, self.pid)
-
-		# debug
-		# print self.headers
-		# print self.body
-
-		# capture modifications to datastream
-		if self.methodName in ['modifyDatastreamByValue','modifyDatastreamByReference']:
-			self._determine_ds()
-			if localConfig.SOLR_AUTOINDEX and self.ds not in localConfig.SKIP_INDEX_DATASTREAMS:
-				self.index_object()
-
-		# capture ingests
-		if self.methodName in ['ingest']:
-			if localConfig.SOLR_AUTOINDEX:
-				self.index_object()
-
-		# capture purge
-		if self.methodName in ['purgeObject']:
-			if localConfig.SOLR_AUTOINDEX:
-				self.purge_object()
-
-
-	def _determine_ds(self):
-		'''
-		Small function to determine which datastream was acted on
-		'''
-		self.ds = [c['@term'] for c in self.categories if c['@scheme'] == 'fedora-types:dsID'][0]
-		print "datastream %s was acted on" % self.ds
-		return self.ds
-
-
-	def index_object(self):
-		return solrIndexer.delay("fedoraConsumerIndex", self.pid)
-
-
-	def purge_object(self):
-		return pruneSolr_worker.delay(None,PID=self.pid)
-
-
-
-
-
-
-
-
-
-
+# ##########################################################################################
+# # setup db
+# ##########################################################################################
+# app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql://%s:%s@localhost/%s' % (localConfig.MYSQL_USERNAME, localConfig.MYSQL_PASSWORD, localConfig.MYSQL_DATABASE ) 
+# app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+# app.config['SQLALCHEMY_POOL_SIZE'] = 1
+# db = SQLAlchemy(app)
